requirements (2).txt
pandas
numpy
scikit-learn
nltk
python-docx
pdfplumber
pywin32
joblib
streamlit